[aml-power][s905d2] Introduce power driver.
This patch introduces an power implementation driver for the amlogic
platform. Currently it is only implemented for the s905d2's CPU
power domain and currently only supports changing the voltage of the
power domain.

The functionality is duplicated from the aml-voltage component of the
aml-thermal driver which is ultimately planned to be removed from the
aml-thermal driver.

Change-Id: I25035d3ea41c1f75b27fba82631f75064b6c0ee5