[kernel][sched] Add ThreadQueueEvents to tracing
Add trace probes to the scheduler which record when a thread enters or
leaves a CPU run queue, or when it becomes active on that CPU.
Also, record the effective affinity mask of the thread at this point
in time.

This allows for an alternate method of computing utilization
statistics, one which can take into account the affinity restrictions
present in the system (simply using Run-Q lengths cannot do this).

Change-Id: I78edfd16fed405f2570d8d648e9d76924d4de00c