[blobfs] Eagerly evict non-pager-backed files
NOTE: this is a special commit to merge with a 0922 build of
Fuchsia. It should be identical to fxrev.dev/435335 except for the
review.

Blobfs retains the pages of a file in memory indefinitely, with the
intent of allowing the kernel to clean up the memory through page
eviction.

Recently, instability during system update has been observed, due to the
system reaching OOM conditions during update. This is because blobs
written during OTA are not pager-backed and thus cannot be subject to
the usual eviction strategy. As a result, all written blobs get stuck
in memory.

This change modifies the handling of the |blobfs.cache-eviction-policy|
kernel commandline knob to only affect pager-backed blobs. This results
in the desired behaviour of blobfs eagerly evicting non-pager-backed
blobs when their last handle closes, while still not eagerly evicting
pager-backed blobs as intended.

Bug: 58097
Tested: Unit test
Tested: Manual test that eviction now correclty clears written blobs
Change-Id: I35507b70d77cb5853da43ee88a24a6bb9b1dc47b